requirements.txt
numpy==1.21.6
pandas==1.3.5
tensorflow==2.8.2
pytorch_lightning==1.6.4
git+https://github.com/facebookresearch/esm.git
tape_proteins
transformers==4.20.1
SentencePiece